+++ src/components/TodoItem.tsx
@@ -38,44 +38,46 @@ export const TodoItem = ({ data, getTodo }: TodoItemPropsType) => {
   };
   return (
     <li key={data.id}>
-      <div>
-        <input id={`checkbox-${data.id}`} type="checkbox" onChange={checkedTodo} checked={data.isCompleted} />
-        {isEdit ? (
-          <input
-            type="text"
-            value={newTodo}
-            onChange={(e) => {
-              setNewTodo(e.target.value);
-            }}
-          />
-        ) : (
-          <label htmlFor={`checkbox-${data.id}`}>{data.todo}</label>
-        )}
-      </div>
-
       {isEdit ? (
-        <div>
-          <button onClick={updatedTodo}>제출</button>
-          <button
-            onClick={() => {
-              setIsEdit(false);
-            }}
-          >
-            취소
-          </button>
-        </div>
+        <>
+          <div>
+            <input
+              type="text"
+              value={newTodo}
+              onChange={(e) => {
+                setNewTodo(e.target.value);
+              }}
+            />
+          </div>
+          <div>
+            <button onClick={updatedTodo}>제출</button>
+            <button
+              onClick={() => {
+                setIsEdit(false);
+              }}
+            >
+              취소
+            </button>
+          </div>
+        </>
       ) : (
-        <div>
-          <button
-            onClick={() => {
-              setNewTodo(data.todo);
-              setIsEdit(true);
-            }}
-          >
-            수정
-          </button>
-          <button onClick={deleteTodo}>삭제</button>
-        </div>
+        <>
+          <div>
+            <input id={`checkbox-${data.id}`} type="checkbox" onChange={checkedTodo} checked={data.isCompleted} />
+            <label htmlFor={`checkbox-${data.id}`}>{data.todo}</label>
+          </div>
+          <div>
+            <button
+              onClick={() => {
+                setNewTodo(data.todo);
+                setIsEdit(true);
+              }}
+            >
+              수정
+            </button>
+            <button onClick={deleteTodo}>삭제</button>
+          </div>
+        </>
       )}
     </li>
   );
